container_util.py can print dictionaries, lists and tuples.
